moved tests in miscellanous
