+++ test/rootchain/blockSubmissions.js
@@ -33,4 +33,28 @@ contract('[RootChain] Block Submissions', async (accounts) => {
         let curr = (await rootchain.lastCommittedBlock.call()).toNumber();
         assert.equal(prev, curr, "Child blocknum incorrectly changed");
     });
+
+    it("Can submit more than one merkle root", async () => {
+        let root1 = web3.sha3("root1").slice(2);
+        let root2 = web3.sha3("root2").slice(2);
+        let roots = root1 + root2;
+
+        let lastCommitedBlock = 0;
+        await rootchain.submitBlock(toHex(roots), [1, 2], 1, {from: authority});
+
+        assert.equal((await rootchain.lastCommittedBlock.call()).toNumber(), 2, "blocknum incremented incorrectly");
+        assert.equal((await rootchain.childChain.call(1))[0], toHex(root1), "mismatch in block root");
+        assert.equal((await rootchain.childChain.call(2))[0], toHex(root2), "mismatch in block root");
+    });
+
+    it("Enforces block number ordering", async () => {
+        let root1 = web3.sha3("root1").slice(2)
+        let root3 = web3.sha3("root3").slice(2)
+
+        await rootchain.submitBlock(toHex(root1), [1], 1);
+        let err;
+        [err] = await catchError(rootchain.submitBlock(toHex(root3), [1], 3));
+        if (!err)
+            assert.fail("Allowed block submission with inconsistent ordering");
+    });
 });
